// frontend/src/interfaces/Controller.ts
interface Dimensions {
  width: number | string;
  height: number | string;
}
// This is for item position
interface Position {
  top: number | string;
  left: number | string;
}
// This is for admin app only
interface Selected {
  selected?: boolean;
}
// This is the base controller
interface BaseController extends Dimensions, Position, Selected {
  id: number | string;
  type: string;
  pageId: number;
  css: string;
  cssClass: string;
  active: boolean;
}
// Below interfaces are for all specific controllers
export interface Image extends BaseController {
  type: 'image';
  srcType: 'local' | 'external';
  value: string;
  objectFit: string;
}
export interface Page extends BaseController {
  type: 'page';
  targetPage: string;
  title: string;
}